frontend/requirements.txt
django-bootstrap3
django-admin-bootstrapped
requests
slumber
django-crispy-forms
django-allauth
